added GPT-OSS 120B (bs=1) MegaKernel on mi355

+++ src/transpiler/plan_tb_swizzle_blackwell.cc
@@ -157,8 +157,7 @@ void Transpiler::get_threadblock_swizzle_plan_blackwell(
       // This rule should be aligned with resolve_tensor_layout.cc
       assert(stensor.dim[meta.innermost_dim] != 1);
       meta.strides[meta.innermost_dim] = 1;
-      size_t cur_stride =
-          static_cast<size_t>(new_num_chunks_in_inner_dim) * chunk_size_num_elems;
+      size_t cur_stride = new_num_chunks_in_inner_dim * chunk_size_num_elems;
       // printf("%u -> %lu\n", num_chunks_in_inner_dim * chunk_size_num_elems,
       // cur_stride);
       for (int i = num_dims - 1; i >= 0; --i) {
